requirements.txt
aiohttp==3.9.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
alembic==1.13.1
attrs==23.2.0
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coverage==7.4.4
dnspython==2.6.1
filelock==3.13.4
Flask==3.0.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
frozenlist==1.4.1
greenlet==3.0.3
idna==3.7
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.3
Mako==1.3.3
MarkupSafe==2.1.5
multidict==6.0.5
mysql-connector-python==8.3.0
mysqlclient==2.2.4
packaging==24.0
passlib==1.7.4
pluggy==1.4.0
py3-validate-email @ git+https://gitea.ksol.io/karolyi/py3-validate-email@3f3fcc241c03e85f0f8466992cded1f715f04d67
PyJWT==2.8.0
pytest==8.1.1
pytest-mock==3.14.0
python-dotenv==1.0.1
requests==2.31.0
SQLAlchemy==2.0.29
twilio==9.0.4
typing_extensions==4.11.0
urllib3==2.2.1
Werkzeug==3.0.3
WTForms==3.1.2
yarl==1.9.4
